adição da busca por id e busca geral dos clientes

+++ main/src/main/java/com/project/app/main/services/ClientService.java
@@ -10,6 +10,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
+import java.util.List;
 
 @Slf4j
 @Service
@@ -56,9 +57,25 @@ public class ClientService {
     }
 
     public String deleteClient(Long clientId){
+        if(clientId == null){
+            throw new BadRequestException("por favor, informe um Id válido");
+        }
         Client client = repository.findById(clientId).
                 orElseThrow( () -> new NotFoundException("não foi possível encontrar um cliente com esse id"));
         repository.delete(client);
         return "client deletado com sucesso!";
     }
+
+    public Client findClientById(Long clientId){
+        if(clientId == null){
+            throw new BadRequestException("por favor, informe um Id válido");
+        }
+        return repository.findById(clientId).
+                orElseThrow( () -> new NotFoundException("não foi possível encontrar um cliente com esse id"));
+    }
+
+    public List<Client> getAllClients(){
+        return repository.findAll();
+    }
+
 }
